backend/requirements_fastapi.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pyyaml
requests
openpyxl
python-dotenv
apscheduler
